+++ src/app/_components/organism/form/LoginForm.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Form } from "./LoginForm.style";
-import Labeled_Input from "../../molecules/Input/Labeled_Input";
+import LabeledInput from "../../molecules/Input/LabeledInput";
 import Button from "../../atoms/button/BlueButton";
 
 export default function LoginForm() {
@@ -14,8 +14,8 @@ export default function LoginForm() {
 
   return (
     <Form>
-      <Labeled_Input label="전화번호, 사용자 이름 또는 이메일" type="text" value={text} onChange={setText} />
-      <Labeled_Input label="비밀번호" type="password" value={password} onChange={setPassword} />
+      <LabeledInput label="전화번호, 사용자 이름 또는 이메일" type="text" value={text} onChange={setText} />
+      <LabeledInput label="비밀번호" type="password" value={password} onChange={setPassword} />
       <Button value="로그인" onClick={buttonHandler} isActive={isActive} />
     </Form>
   );
